% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ctrl_r1r2_ratio}
\alias{ctrl_r1r2_ratio}
\title{ctrl_r1r2_ratio}
\usage{
ctrl_r1r2_ratio(df)
}
\arguments{
\item{df}{data frame in  various KO objectives}
}
\value{

}
\description{
ctrl_r1r2_ratio
}
\examples{
ctrl_r1r2_ratio(df)
}
