% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_report}
\alias{error_report}
\title{error_report}
\usage{
error_report(df)
}
\arguments{
\item{df}{data frame}
}
\value{

}
\description{
error_report
}
\examples{
error_report <- error_report(BE1)
}
