% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KO1_r1r2_ratio}
\alias{KO1_r1r2_ratio}
\title{KO1_r1r2_ratio}
\usage{
KO1_r1r2_ratio(df)
}
\arguments{
\item{df}{data frame in  KO-1 objectives}
}
\value{
data
}
\description{
KO1_r1r2_ratio
}
\examples{
KO1_r1r2_ratio(KO1)
}
