% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KO_r1r2_ratio}
\alias{KO_r1r2_ratio}
\title{KO_r1r2_ratio}
\usage{
KO_r1r2_ratio(df)
}
\arguments{
\item{df}{data frame in  various KO objectives}
}
\value{

}
\description{
KO_r1r2_ratio
}
\examples{
KO_r1r2_ratio(df)
}
