% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{newdf}
\alias{newdf}
\title{newdf}
\usage{
newdf(KOratio, ctrlratio)
}
\arguments{
\item{KOratio}{data frame from KO objectives}

\item{ctrlratio}{data frame from ctrl objectives}
}
\value{

}
\description{
newdf
}
\examples{
newdf <-newdf(KOratio, ctrlratio)
}
