% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{r1r2_ratio}
\alias{r1r2_ratio}
\title{r1r2_ratio}
\usage{
r1r2_ratio(df)
}
\arguments{
\item{df}{data frame in  various ctrl or KO objectives}
}
\value{

}
\description{
r1r2_ratio
}
\examples{
r1r2_ratio("df")
}
